touchpad: Add button debouncing
It is possible for touchpad physical buttons to exhibit the
same bouncing behavior that can be seen on regular pointer
device buttons. In fact, any mechanical switch is likely to
have some level of bouncing that should be filtered.

So, send touchpad button events through the same debounce filter
logic that is currently being used by fallback_dispatch.

Signed-off-by: Jacob Moroni <mail@jakemoroni.com>