touchpad: implement hold-and-tap
This feature allows a user, as the name suggests, to hold fingers on a
touchpad while tapping with additional fingers to produce the same effect as
without the holding fingers.
The holding fingers may even have moved, pressed the clickpad button (unless
it is still down at the time of the tap), or be participating in a
tap-and-drag, though in the latter case a tap with the same number of fingers
as was used to start the tap-and-drag is ignored, just like a click from
another mouse-like device would be.

Signed-off-by: satrmb <10471-satrmb@users.noreply.gitlab.freedesktop.org>