touchpad: if hold-and-tap is on, treat half-released taps as lesser taps
The expectation is that the user means the remaining fingers of the tap
to be holding fingers which just happened to be put down at roughly
the same time as the released tapping fingers.

This necessitates handling a tap with four or more fingers, because only
three of them might get lifted, making it actually a three-finger tap
and at least one more holding finger.

Though not strictly necessary, these states can also be reached
if hold-and-tap is off: it may catch a many-finger-tap to end drag-lock,
or a three-finger-tap joined by a palm, as demonstrated in the
previously existing test case touchpad_tap_palm_on_touch_4.

Signed-off-by: satrmb <10471-satrmb@users.noreply.gitlab.freedesktop.org>